test: hanoi: middleware: use connexion and openapi
Tests for the OpenAPI / Connexxion middleware.

Unfortunately, it does not seem possible to run an integration test with pytest
and the openapi / connexion app at this time as the server is not reachable within
the GitHub Actions & pytest environment - each python requests call is unable to
connect. The test results were fine on my machine though, so it would really just
be a matter of taking the time to drill down and figure out what is up with the
GitHub Actions environment. It's possible that it might be restrictive in terms of
which capabilities are exposed inside of a container, etc.

Results on my machine:

$ PYTHONPATH=$PWD/src pytest tests/app_test.py
================================= test session starts =================================
platform linux -- Python 3.8.5, pytest-6.1.1, py-1.9.0, pluggy-0.13.1
rootdir: /home/cfriedt/workspace/hanoi
plugins: subtests-0.2.1, schemathesis-1.1.0, cov-2.10.1, hypothesis-5.41.4, xdist-1.31.0, forked-1.1.3
collected 13 items

tests/app_test.py ...s.........                                                 [100%]

============================ 12 passed, 1 skipped in 0.95s ============================

Fixes #2

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>